+++ CHANGELOG.md
@@ -322,3 +322,4 @@
 - 2024-03-07: Async follow up, ready pool, mutate, main, +4
 - 2024-03-08: Err deal, frontier, contacts, icp, +5
 - 2024-03-09: Integ find email, task, icp, mutate, +6
+- 2024-03-10: Shutdown sender, contacts, follow up, ready pool, +2